requirements.txt
# ERP Purchase Order System - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
# psycopg2-binary==2.9.9  # Uncomment if using PostgreSQL
aiosqlite==0.19.0  # For SQLite async support

# Authentication
python-jose[cryptography]==3.3.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# Data Validation
pydantic[email]==2.6.0
email-validator==2.1.0

# Environment Variables
python-dotenv==1.0.1

# HTTP Client (for AI descriptions)
httpx==0.26.0

# CORS and Security
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# Security
bleach==6.1.0

# Development
watchfiles==0.21.0
